current/miniproj-week5and6-DBfullOOP/MP-python-docker/app/requirements.txt
cffi==1.15
cryptography==35.0.0
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.19.2
six==1.16.0
tabulate==0.8.9